.ipynb_checkpoints/requirements-checkpoint.txt
# UI
streamlit

# Core AI stack
langchain
transformers
sentence-transformers
faiss-cpu
torch
langchain-community

# Web scraping
beautifulsoup4
requests

# Environment variable loader
python-dotenv
